VIVO-443 Use language-specific templates to create email
For some reason, the password-reset template was too long to do proper substitution on the last line. Removing some of the intermediate content fixed the problem, but without appropriate content. I had been unhappy with the @@file approach anyway, so I went back to language-specific templates for each type of message.